remove ExprNode
