+++ lang/test/org/partiql/lang/planner/memorydb/MemoryTable.kt
@@ -1,6 +1,5 @@
 package org.partiql.lang.planner.memorydb
 
-import org.partiql.lang.ast.DeleteOp.name
 import org.partiql.lang.eval.BindingCase
 import org.partiql.lang.eval.BindingName
 import org.partiql.lang.eval.DEFAULT_COMPARATOR
@@ -45,7 +44,7 @@ class MemoryTable(
         val primaryKeyExprValue = row.extractPrimaryKey()
 
         if (rows.containsKey(primaryKeyExprValue)) {
-            error("Table '$name' already contains a row with the specified primary key ")
+            error("Table '${this.metadata.name}' already contains a row with the specified primary key ")
         } else {
             // We have to detatch the ExprValue from any lazily evaluated query that may get invoked
             // whenever the value is accessed.  To do this we convert to Ion, which forces full materialization,
